# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vertex AI Mailer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns:
    get:
      operationId: getCampaigns
      tags: [campaigns]
      summary: List all campaigns for current user
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [drafted, pending, failed, completed]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Campaigns list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCampaigns"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a new campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get campaign by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Campaign updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /campaigns/{id}/generate-drafts:
    post:
      operationId: generateCampaignDrafts
      tags: [campaigns]
      summary: Generate Gmail drafts for all leads in a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateDraftsInput"
      responses:
        "200":
          description: Draft generation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DraftGenerationResult"

  # ─── LEADS ──────────────────────────────────────────────────────────────────
...
  /templates:
    get:
      operationId: getTemplates
      tags: [templates]
      summary: List email templates for current user
      responses:
        "200":
          description: Templates list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmailTemplate"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create email template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailTemplateInput"
      responses:
        "201":
          description: Template created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get template by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
    patch:
      operationId: updateTemplate
      tags: [templates]
      summary: Update template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailTemplateUpdate"
      responses:
        "200":
          description: Template updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /templates/{id}/duplicate:
    post:
      operationId: duplicateTemplate
      tags: [templates]
      summary: Duplicate a template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"

  # ─── AI ─────────────────────────────────────────────────────────────────────
...
  /drafts:
    get:
      operationId: getDrafts
      tags: [drafts]
      summary: List all draft records
      parameters:
        - name: campaignId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, success, failed]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Drafts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedDrafts"
  /drafts/{id}:
    get:
      operationId: getDraft
      tags: [drafts]
      summary: Get draft record by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Draft record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DraftRecord"

  # ─── UPLOADS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ─── AUTH ────────────────────────────────────────────────────────────────
...
    Campaign:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [drafted, pending, failed, completed]
        templateId:
          type: ["integer", "null"]
        totalLeads:
          type: integer
        draftedCount:
          type: integer
        failedCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, name, status, totalLeads, draftedCount, failedCount, createdAt, updatedAt]
    CampaignInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        templateId:
          type: ["integer", "null"]
      required: [name]
    CampaignUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        status:
          type: string
          enum: [drafted, pending, failed, completed]
        templateId:
          type: ["integer", "null"]
    PaginatedCampaigns:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Campaign"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
    GenerateDraftsInput:
      type: object
      properties:
        templateId:
          type: integer
        tone:
          type: string
          enum: [professional, friendly, sales, followup, urgent]
          default: professional
        customPrompt:
          type: ["string", "null"]
      required: [templateId]
    DraftGenerationResult:
      type: object
      properties:
        total:
          type: integer
        succeeded:
          type: integer
        failed:
          type: integer
        errors:
          type: array
          items:
            type: string
      required: [total, succeeded, failed, errors]

    # ─── LEADS ───────────────────────────────────────────────────────────────
...
    EmailTemplate:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        subject:
          type: string
        body:
          type: string
        isDefault:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, name, subject, body, isDefault, createdAt, updatedAt]
    EmailTemplateInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        subject:
          type: string
          minLength: 1
        body:
          type: string
          minLength: 1
        isDefault:
          type: boolean
      required: [name, subject, body]
    EmailTemplateUpdate:
      type: object
      properties:
        name:
          type: string
        subject:
          type: string
        body:
          type: string
        isDefault:
          type: boolean

    # ─── AI ──────────────────────────────────────────────────────────────────
...
    DraftRecord:
      type: object
      properties:
        id:
          type: integer
        campaignId:
          type: ["integer", "null"]
        leadId:
          type: integer
        userId:
          type: integer
        gmailDraftId:
          type: ["string", "null"]
        subject:
          type: string
        body:
          type: string
        status:
          type: string
          enum: [pending, success, failed]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, leadId, userId, subject, body, status, createdAt]
    PaginatedDrafts:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/DraftRecord"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]

    # ─── UPLOADS ─────────────────────────────────────────────────────────────
...