# lib/api-zod/src/generated/types/draftRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vertex AI Mailer API
 * OpenAPI spec version: 0.1.0
 */
import type { DraftRecordStatus } from './draftRecordStatus';

export interface DraftRecord {
  id: number;
  /** @nullable */
  campaignId?: number | null;
  leadId: number;
  userId: number;
  /** @nullable */
  gmailDraftId?: string | null;
  subject: string;
  body: string;
  status: DraftRecordStatus;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
}